pyproject.toml
[project]
name = "ds-rpc-01"
version = "0.1.0"
description = "Starter project for the RPC-01: Internal Chatbot with Role Based Access Control"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=0.6.3",
    "cohere>=5.15.0",
    "fastapi[standard]>=0.115.12",
    "ipykernel>=6.29.5",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-google-genai>=2.1.5",
    "langchain-groq>=0.3.2",
    "langchain-huggingface>=0.3.0",
    "langgraph>=0.4.8",
    "markdown>=3.8",
    "python-dotenv>=1.1.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=4.1.0",
    "unstructured[markdown]>=0.17.2",
]
